requirements.txt
apache-airflow==2.6.0
langchain==0.0.170
matplotlib==3.7.1
numpy==1.24.3
oauth2client==4.1.3
openai==0.27.6
pandas==2.0.1
pandas_gbq==0.19.2
plotly==5.14.1
protobuf==3.20.3
python-dotenv==1.0.0
python_dateutil==2.8.2
regex==2023.5.5
Requests==2.30.0
scholarly==1.7.11
selenium==4.9.1
streamlit==1.22.0
urllib3==1.26.15
regex
